Stage 1 commit of trying to implement containers + type flows, with the ability to have nested containers (List<List<int>>) in mind. This doesn't compile, but I want a point I can revert back to before I continue.

+++ SonezakiMasaki/Exceptions/UnrecognizedContainerException.cs
@@ -7,12 +7,12 @@ namespace SonezakiMasaki.Exceptions
 {
     public sealed class UnrecognizedContainerException : SonezakiMasakiException
     {
-        internal UnrecognizedContainerException( byte containerId )
+        internal UnrecognizedContainerException( int containerId )
             : base( $"There is no container with the encountered ID of {containerId}." )
         {
             ContainerId = containerId;
         }
 
-        public byte ContainerId { get; }
+        public int ContainerId { get; }
     }
 }
